--- a/mujoco_warp/_src/collision_gjk.py
+++ b/mujoco_warp/_src/collision_gjk.py
@@ -584,7 +584,7 @@ def _S1D(s1: wp.vec3, s2: wp.vec3):
 
 
 @wp.func
-def _gjk(
+def gjk(
   # In:
   tolerance: float,
   gjk_iterations: int,
@@ -2095,7 +2095,7 @@ def ccd(
   # special handling for sphere and capsule (shrink to point and line respectively)
   if margin1 + margin2 > 0.0:
     cutoff += margin1 + margin2
-    result = _gjk(tolerance, gjk_iterations, geom1, geom2, x_1, x_2, geomtype1, geomtype2, cutoff, True)
+    result = gjk(tolerance, gjk_iterations, geom1, geom2, x_1, x_2, geomtype1, geomtype2, cutoff, True)
 
     # shallow penetration, inflate contact
     if result.dist > tolerance:
@@ -2111,7 +2111,7 @@ def ccd(
     # deep penetration, reset initial conditions and rerun GJK + EPA
     cutoff -= margin1 + margin2
 
-  result = _gjk(tolerance, gjk_iterations, geom1, geom2, x_1, x_2, geomtype1, geomtype2, cutoff, False)
+  result = gjk(tolerance, gjk_iterations, geom1, geom2, x_1, x_2, geomtype1, geomtype2, cutoff, False)
 
   # no penetration depth to recover
   if result.dist > tolerance or result.dim < 2: