**benchmark/apptronik_apollo/scene_hfield.xml**
```
mjwarp-testspeed benchmark/apptronik_apollo/scene_hfield.xml --nconmax=3000000 --njmax=256
```

this pr:
```
Summary for 8192 parallel rollouts

Total JIT time: 0.59 s
Total simulation time: 24.17 s
Total steps per second: 338,974
Total realtime factor: 1,694.87 x
Total time per step: 2950.08 ns
Total converged worlds: 8192 / 8192
```

main:
```
Summary for 8192 parallel rollouts

Total JIT time: 0.57 s
Total simulation time: 25.64 s
Total steps per second: 319,439
Total realtime factor: 1,597.20 x
Total time per step: 3130.49 ns
Total converged worlds: 8192 / 8192
```

---

reduce nconmax by 10x: 3_000_000 -> 300_000

```
mjwarp-testspeed benchmark/apptronik_apollo/scene_hfield.xml --nconmax=300000 --njmax=256
```

this pr:
```
Summary for 8192 parallel rollouts

Total JIT time: 0.58 s
Total simulation time: 22.91 s
Total steps per second: 357,610
Total realtime factor: 1,788.05 x
Total time per step: 2796.34 ns
Total converged worlds: 8192 / 8192
```

main:
```
Rolling out 1000 steps at dt = 0.005...
ncollision overflow - please increase nconmax to 440544
```