@thowell 2025-09-05
the pr has been refactored and the pr description has been updated with new timing results.

the primary change is removing the midphase routine and replacing it with a narrowphase routine.