@thowell 2025-09-05
updated pr description with timing results for this pr after adding `MJ_MAXCONPAIR` check in `ccd_hf_kernel`